requirements.txt
beautifulsoup4==4.13.5
google_api_python_client==2.179.0
langchain_community==0.3.29
langchain_google_genai==2.1.10
langchain_openai==0.3.32
langchain_text_splitters==0.3.9
pandas==2.3.2
protobuf==6.32.0
python-dotenv==1.1.1
Requests==2.32.5
supabase==2.18.1
pymupdf==1.21.1
